requirements.txt
streamlit==1.11.0
pandas==1.4.2
numpy==1.21.5
altair==4.2.0
plotly-express==0.4.1
matplotlib==3.5.1